Implemented get tokens helper for manage tokens scene
