+++ src/components/common/ManageTokens/index.js
@@ -1,3 +1,4 @@
 // @flow
 
 export { default as TokenRow } from './TokenRow'
+export * from './helpers'
