+++ src/components/scenes/ManageTokensScene.js
@@ -13,7 +13,7 @@ import s from '../../locales/strings.js'
 import { type RootState } from '../../types/reduxTypes.js'
 import type { CustomTokenInfo, GuiWallet } from '../../types/types.js'
 import * as UTILS from '../../util/utils'
-import { TokenRow } from '../common/ManageTokens'
+import { getTokens, TokenRow } from '../common/ManageTokens'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext'
 import SceneFooter from '../themed/SceneFooter'
@@ -75,32 +75,13 @@ class ManageTokensScene extends React.Component<ManageTokensProps, State> {
   render() {
     const { metaTokens, currencyCode } = this.props.guiWallet
     const { manageTokensPending } = this.props
-    let accountMetaTokenInfo = []
-    const specialCurrencyInfo = getSpecialCurrencyInfo(currencyCode)
-    // this will need refactoring later
-    if (specialCurrencyInfo.isCustomTokensSupported) {
-      accountMetaTokenInfo = [...this.props.settingsCustomTokens]
-    }
-    const filteredTokenInfo = accountMetaTokenInfo.filter(token => {
-      return token.walletType === this.props.guiWallet.type || token.walletType === undefined
-    })
-    const combinedTokenInfo = UTILS.mergeTokensRemoveInvisible(metaTokens, filteredTokenInfo)
 
-    const sortedTokenInfo = combinedTokenInfo.sort((a, b) => {
-      if (a.currencyCode < b.currencyCode) return -1
-      if (a === b) return 0
-      return 1
+    const tokens = getTokens({
+      metaTokens,
+      customTokens: this.props.settingsCustomTokens,
+      currencyCode,
+      guiWalletType: this.props.guiWallet.type
     })
-
-    // put preferred tokens at the top
-    for (const cc of PREFERRED_TOKENS) {
-      const idx = sortedTokenInfo.findIndex(e => e.currencyCode === cc)
-      if (idx > -1) {
-        const tokenInfo = sortedTokenInfo[idx]
-        sortedTokenInfo.splice(idx, 1)
-        sortedTokenInfo.unshift(tokenInfo)
-      }
-    }
 
     const { theme } = this.props
 
@@ -113,7 +94,7 @@ class ManageTokensScene extends React.Component<ManageTokensProps, State> {
             <View style={styles.tokensArea}>
               <FlatList
                 keyExtractor={item => item.currencyCode}
-                data={sortedTokenInfo}
+                data={tokens}
                 renderItem={metaToken => (
                   <TokenRow
                     goToEditTokenScene={this.goToEditTokenScene}
